Move va-subdomain
