+++ src/platform/site-wide/user-nav/components/HelpMenu.jsx
@@ -4,7 +4,7 @@ import React from 'react';
 import DropDownPanel from '@department-of-veterans-affairs/formation-react/DropDownPanel';
 import IconHelp from '@department-of-veterans-affairs/formation-react/IconHelp';
 
-import isVATeamSiteSubdomain from '../../../brand-consolidation/va-subdomain';
+import isVATeamSiteSubdomain from '../../../utilities/environment/va-subdomain';
 import facilityLocatorManifest from '../../../../applications/facility-locator/manifest';
 
 const FACILITY_LOCATOR_URL = facilityLocatorManifest.rootUrl;
